Code Refactoring: Move delta-sharing-client out as a separate directory (#334)
* client lib

* fix scalastyle

* fix

* fix

* fix scalastyle

* fix tests

* move JsonPredicate/Opconverter back to spark

* fix upgrade
