+++ client/src/main/resources/META-INF/services/org.apache.hadoop.fs.FileSystem
@@ -1,0 +1,1 @@
+io.delta.sharing.client.DeltaSharingFileSystem
