+++ spark/src/main/resources/META-INF/services/org.apache.hadoop.fs.FileSystem
@@ -1,1 +1,1 @@
-io.delta.sharing.spark.DeltaSharingFileSystem
+io.delta.sharing.client.DeltaSharingFileSystem
