--- spark/src/main/scala/io/delta/sharing/spark/util/JsonUtils.scala
@@ -1,40 +1,0 @@
-/*
- * Copyright (2021) The Delta Lake Project Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.delta.sharing.spark.util
-
-import com.fasterxml.jackson.annotation.JsonInclude.Include
-import com.fasterxml.jackson.databind.{DeserializationFeature, ObjectMapper}
-import com.fasterxml.jackson.module.scala.{DefaultScalaModule, ScalaObjectMapper}
-
-private[sharing] object JsonUtils {
-  /** Used to convert between classes and JSON. */
-  lazy val mapper = {
-    val _mapper = new ObjectMapper with ScalaObjectMapper
-    _mapper.setSerializationInclusion(Include.NON_ABSENT)
-    _mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
-    _mapper.registerModule(DefaultScalaModule)
-    _mapper
-  }
-
-  def toJson[T: Manifest](obj: T): String = {
-    mapper.writeValueAsString(obj)
-  }
-
-  def fromJson[T: Manifest](json: String): T = {
-    mapper.readValue[T](json)
-  }
-}
